deepset-simclr/requirements.txt
pyyaml
dacite
numpy
torch
torchvision
lightly
pillow
wandb
